+++ AI/requirements.txt
@@ -1,0 +1,4 @@
+SpeechRecognition
+google-generativeai
+hackerbot
+AI_utils
